Refactor: Restructuring the DTO directory
