+++ src/main/java/code/rice/bowl/spaghetti/controller/TopicController.java
@@ -1,7 +1,7 @@
 package code.rice.bowl.spaghetti.controller;
 
-import code.rice.bowl.spaghetti.dto.TopicDto;
-import code.rice.bowl.spaghetti.dto.response.TopicResponse;
+import code.rice.bowl.spaghetti.dto.topic.TopicDto;
+import code.rice.bowl.spaghetti.dto.topic.TopicResponse;
 import code.rice.bowl.spaghetti.service.TopicService;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
