+++ src/main/java/code/rice/bowl/spaghetti/dto/level/LevelDto.java
@@ -1,4 +1,4 @@
-package code.rice.bowl.spaghetti.dto;
+package code.rice.bowl.spaghetti.dto.level;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
