+++ src/main/java/code/rice/bowl/spaghetti/dto/user/CurrentUserResponse.java
@@ -1,4 +1,4 @@
-package code.rice.bowl.spaghetti.dto.response;
+package code.rice.bowl.spaghetti.dto.user;
 
 
 import lombok.Builder;
