+++ src/main/java/code/rice/bowl/spaghetti/dto/userproblemhistory/UserProblemHistorySummaryResponse.java
@@ -1,4 +1,4 @@
-package code.rice.bowl.spaghetti.dto.response;
+package code.rice.bowl.spaghetti.dto.userproblemhistory;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
