+++ src/main/java/code/rice/bowl/spaghetti/dto/userprogress/UserProgressResponse.java
@@ -1,4 +1,4 @@
-package code.rice.bowl.spaghetti.dto.response;
+package code.rice.bowl.spaghetti.dto.userprogress;
 
 
 import lombok.Builder;
