+++ src/main/java/code/rice/bowl/spaghetti/dto/usertopic/UserTopicResponse.java
@@ -1,4 +1,4 @@
-package code.rice.bowl.spaghetti.dto.response;
+package code.rice.bowl.spaghetti.dto.usertopic;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
